fix(widgets): share provider cost merge
