+++ CodexBarMobile/AppStoreMetadata/1.16.0/ja/release_notes.txt
@@ -4,9 +4,9 @@
 * ホーム画面ウィジェット — 小・中・大サイズ、または iPad の特大サイズの CodexBar ウィジェットで、プロバイダ使用量、今日のコスト、同期状態をすばやく確認できます。
 * ライト、ダーク、ホーム画面のTinted表示 — ウィジェットは固定の暗い背景やカラフルなダッシュボード風ではなく、システムの外観に追従します。
 * ウィジェットのカラースタイル — ネイティブな単色表示の Mono、または同じクリーンなレイアウトに控えめなアクセント色を加える Colorful を選べます。
-* Today Cost の調整 — 小サイズではトークン使用量を表示し、中サイズではコストとトークンを見やすく左右に配置します。プロバイダー行ではアカウントプラン名を表示しません。
+* Today Cost の調整 — ウィジェットは Cost ページと同じ統合済みコスト合計を使い、小サイズではトークン使用量を表示し、中サイズではコストとトークンを見やすく左右に配置します。プロバイダー行ではアカウントプラン名を表示しません。
 * 設定可能な表示 — 各ウィジェットで Overview、Provider Focus、Today Cost、Sync Health を選べます。ホーム画面で編集した後も、選択した表示が保持されます。
-* 実際の同期データ — ウィジェットは CodexBar の CloudKit 同期データを読み取り、必要に応じて古い iCloud スナップショットへフォールバックします。
+* 実際の同期データ — ウィジェットは複数デバイスのローカル CLI 支出を含む CodexBar の共有 CloudKit 同期マージを読み取り、必要に応じて古い iCloud スナップショットへフォールバックします。
 * コスト合計 — Cost ページは同期済みプロバイダのサマリーを照合し、日別履歴が不完全でも支出を過小表示しません。
 
 この更新では App Group の設定も CloudKit schema の変更も不要です。
