+++ CodexBarMobile/AppStoreMetadata/1.16.0/zh-Hans/release_notes.txt
@@ -4,9 +4,9 @@
 * 桌面小组件 —— 添加小号、中号、大号或 iPad 超大号 CodexBar 小组件，一眼查看提供商用量、今日成本和同步健康状态。
 * 浅色、深色和主屏幕着色外观 —— 小组件会跟随系统外观，不再使用固定深色或彩色仪表盘风格。
 * 小组件颜色风格 —— 可以选择原生单色感的「纯色」，也可以选择在同一套干净布局上加入克制强调色的「多彩」。
-* Today Cost 细节优化 —— 小号组件现在显示 token 用量，中号组件采用更清晰的金额 / token 左右布局，Provider 行不再显示账号套餐文案。
+* Today Cost 细节优化 —— 小组件现在使用和 Cost 页一致的合并成本总额，小号组件显示 token 用量，中号组件采用更清晰的金额 / token 左右布局，Provider 行不再显示账号套餐文案。
 * 可配置视图 —— 每个小组件都可以选择概览、提供商焦点、今日成本或同步健康；在主屏幕编辑后会保持并显示选中的视图。
-* 真实同步数据 —— 小组件读取 CodexBar CloudKit 同步数据，并在需要时回退到旧版 iCloud 快照。
+* 真实同步数据 —— 小组件读取 CodexBar 共享 CloudKit 同步融合结果，包含多设备本地 CLI 花费，并在需要时回退到旧版 iCloud 快照。
 * 成本总额 —— Cost 页现在会校验已同步的 provider 汇总，daily 历史不完整时不会再低估花费。
 
 本次更新不需要 App Group 设置，也没有 CloudKit schema 变更。
