+++ CodexBarMobile/AppStoreMetadata/1.16.0/zh-Hant/release_notes.txt
@@ -4,9 +4,9 @@
 * 主畫面小工具 —— 加入小型、中型、大型或 iPad 超大型 CodexBar 小工具，一眼查看提供商用量、今日成本和同步健康狀態。
 * 淺色、深色和主畫面著色外觀 —— 小工具會跟隨系統外觀，不再使用固定深色或彩色儀表板風格。
 * 小工具顏色風格 —— 可以選擇原生單色感的「純色」，也可以選擇在同一套乾淨版面上加入克制強調色的「多彩」。
-* Today Cost 細節最佳化 —— 小型小工具現在顯示 token 用量，中型小工具採用更清晰的金額 / token 左右佈局，Provider 行不再顯示帳號方案文案。
+* Today Cost 細節最佳化 —— 小工具現在使用和 Cost 頁一致的合併成本總額，小型小工具顯示 token 用量，中型小工具採用更清晰的金額 / token 左右佈局，Provider 行不再顯示帳號方案文案。
 * 可設定檢視 —— 每個小工具都可以選擇概覽、提供商焦點、今日成本或同步健康；在主畫面編輯後會保留並顯示選取的檢視。
-* 真實同步資料 —— 小工具讀取 CodexBar CloudKit 同步資料，並在需要時回退到舊版 iCloud 快照。
+* 真實同步資料 —— 小工具讀取 CodexBar 共享 CloudKit 同步融合結果，包含多裝置本機 CLI 花費，並在需要時回退到舊版 iCloud 快照。
 * 成本總額 —— Cost 頁現在會校驗已同步的 provider 彙總，daily 歷史不完整時不會再低估花費。
 
 本次更新不需要 App Group 設定，也沒有 CloudKit schema 變更。
