+++ CodexBarMobile/project.yml
@@ -35,7 +35,7 @@ targets:
         PRODUCT_BUNDLE_IDENTIFIER: com.o1xhack.codexbar.sync
         GENERATE_INFOPLIST_FILE: true
         MARKETING_VERSION: "1.16.0"
-        CURRENT_PROJECT_VERSION: "177"
+        CURRENT_PROJECT_VERSION: "178"
 
   CodexBarMobilePushExtension:
     type: app-extension
@@ -56,7 +56,7 @@ targets:
       base:
         PRODUCT_BUNDLE_IDENTIFIER: com.o1xhack.codexbar.mobile.pushextension
         MARKETING_VERSION: "1.16.0"
-        CURRENT_PROJECT_VERSION: "177"
+        CURRENT_PROJECT_VERSION: "178"
         DEVELOPMENT_TEAM: 3TUERHN53E
         INFOPLIST_FILE: CodexBarMobilePushExtension/Info.plist
         CODE_SIGN_ENTITLEMENTS: CodexBarMobilePushExtension/PushExtension.entitlements
@@ -82,7 +82,7 @@ targets:
       base:
         PRODUCT_BUNDLE_IDENTIFIER: com.o1xhack.codexbar.mobile.widgets
         MARKETING_VERSION: "1.16.0"
-        CURRENT_PROJECT_VERSION: "177"
+        CURRENT_PROJECT_VERSION: "178"
         DEVELOPMENT_TEAM: 3TUERHN53E
         INFOPLIST_FILE: CodexBarMobileWidgets/Info.plist
         CODE_SIGN_ENTITLEMENTS: CodexBarMobileWidgets/WidgetExtension.entitlements
@@ -107,7 +107,7 @@ targets:
       base:
         PRODUCT_BUNDLE_IDENTIFIER: com.o1xhack.codexbar.mobile
         MARKETING_VERSION: "1.16.0"
-        CURRENT_PROJECT_VERSION: "177"
+        CURRENT_PROJECT_VERSION: "178"
         DEVELOPMENT_TEAM: 3TUERHN53E
         INFOPLIST_FILE: CodexBarMobile/Info.plist
         CODE_SIGN_ENTITLEMENTS: CodexBarMobile/CodexBarMobile.entitlements
